+++ main.py
@@ -1,5 +1,5 @@
-from src.models import Product, Category
-
+from src.product import Product
+from src.category import Category
 
 if __name__ == "__main__":
     product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
